fix: filter Lancamento find by status
