+++ src/lancamento/repositories/lancamento-history.repository.ts
@@ -1,14 +1,13 @@
 import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
+import { compactQuery } from 'src/utils/console-utils';
 import { SqlDateOperator } from 'src/utils/sql/interfaces/sql-date-operator.interface';
 import { EntityCondition } from 'src/utils/types/entity-condition.type';
 import { dateMonthToHumanMonth } from 'src/utils/types/human-month.type';
 import { Between, DeepPartial, DeleteResult, FindManyOptions, FindOneOptions, FindOptionsWhere, QueryRunner, Repository, SaveOptions, UpdateResult } from 'typeorm';
 import { QueryDeepPartialEntity } from 'typeorm/query-builder/QueryPartialEntity';
-import { LancamentoAutorizacaoHistory } from './entities/lancamento-autorizacao-history.entity';
-import { LancamentoAutorizacao } from './entities/lancamento-autorizacao.entity';
-import { LancamentoHistory } from './entities/lancamento-history.entity';
-import { Lancamento } from './entities/lancamento.entity';
+import { LancamentoHistory } from '../entities/lancamento-history.entity';
+import { Lancamento } from '../entities/lancamento.entity';
 
 export interface LancamentoFindWhere {
   detalheA?: { id: number[] };
@@ -16,39 +15,32 @@ export interface LancamentoFindWhere {
 }
 
 @Injectable()
-export class LancamentoAutorizacaoHistoryRepository {
+export class LancamentoHistoryRepository {
   constructor(
-    @InjectRepository(LancamentoAutorizacaoHistory)
-    private readonly lancamentoAutorizacaoHistoryRepository: Repository<LancamentoAutorizacaoHistory>,
-    @InjectRepository(LancamentoAutorizacao)
-    private readonly lancamentoAutorizacaoRepository: Repository<LancamentoAutorizacao>,
+    @InjectRepository(LancamentoHistory)
+    private readonly lancamentoHistoryRepository: Repository<LancamentoHistory>,
   ) {}
 
-  create(entityLike: DeepPartial<LancamentoAutorizacaoHistory>): LancamentoAutorizacaoHistory {
-    return this.lancamentoAutorizacaoHistoryRepository.create(entityLike);
+  create(entityLike: DeepPartial<LancamentoHistory>): LancamentoHistory {
+    return this.lancamentoHistoryRepository.create(entityLike);
   }
 
-  async createBackup(lancamentoHistory: LancamentoHistory, lancamento: Lancamento): Promise<LancamentoAutorizacaoHistory[]> {
-    const autorizacoes = await this.lancamentoAutorizacaoRepository.find({ where: { lancamento: { id: lancamento.id } } });
-    const savedMany: LancamentoAutorizacaoHistory[] = [];
-    for (const autorizacao of autorizacoes) {
-      const autorizacaoHistory = LancamentoAutorizacaoHistory.fromLancamentoAutorizacao(autorizacao, lancamentoHistory);
-      const saved = await this.lancamentoAutorizacaoHistoryRepository.save(autorizacaoHistory);
-      savedMany.push(saved);
-    }
-    return savedMany;
+  async createBackup(lancamento: Lancamento): Promise<LancamentoHistory> {
+    const lancamentoHistory = LancamentoHistory.fromLancamento(lancamento, false);
+    const saved = await this.lancamentoHistoryRepository.save(lancamentoHistory);
+    return saved;
   }
 
-  save(entity: DeepPartial<LancamentoAutorizacaoHistory>, options?: SaveOptions): Promise<LancamentoAutorizacaoHistory> {
-    return this.lancamentoAutorizacaoHistoryRepository.save(entity, options);
+  save(entity: DeepPartial<LancamentoHistory>, options?: SaveOptions): Promise<LancamentoHistory> {
+    return this.lancamentoHistoryRepository.save(entity, options);
   }
 
-  update(criteria: FindOptionsWhere<LancamentoAutorizacaoHistory>, partialEntity: QueryDeepPartialEntity<LancamentoAutorizacaoHistory>, queryRunner?: QueryRunner): Promise<UpdateResult> {
-    return (queryRunner?.manager?.getRepository(Lancamento) || this.lancamentoAutorizacaoHistoryRepository).update(criteria, partialEntity);
+  update(criteria: FindOptionsWhere<LancamentoHistory>, partialEntity: QueryDeepPartialEntity<LancamentoHistory>, queryRunner?: QueryRunner): Promise<UpdateResult> {
+    return (queryRunner?.manager?.getRepository(Lancamento) || this.lancamentoHistoryRepository).update(criteria, partialEntity);
   }
 
-  async findOne(options: FindOneOptions<LancamentoAutorizacaoHistory>): Promise<LancamentoAutorizacaoHistory | null> {
-    let qb = this.lancamentoAutorizacaoHistoryRepository
+  async findOne(options: FindOneOptions<LancamentoHistory>): Promise<LancamentoHistory | null> {
+    let qb = this.lancamentoHistoryRepository
       .createQueryBuilder('lancamento') //
       .leftJoinAndSelect('lancamento.autorizacoes', 'autorizacoes')
       .leftJoinAndSelect('lancamento.autor', 'autor')
@@ -65,7 +57,7 @@ export class LancamentoAutorizacaoHistoryRepository {
     return await qb.getOne();
   }
 
-  async getOne(options: FindOneOptions<LancamentoAutorizacaoHistory>): Promise<LancamentoAutorizacaoHistory> {
+  async getOne(options: FindOneOptions<LancamentoHistory>): Promise<LancamentoHistory> {
     const found = await this.findOne(options);
     if (!found) {
       throw new HttpException('Lancamento não encontrado', HttpStatus.NOT_FOUND);
@@ -73,9 +65,9 @@ export class LancamentoAutorizacaoHistoryRepository {
     return found;
   }
 
-  async findMany(options?: FindManyOptions<LancamentoAutorizacaoHistory> | undefined, andWhere?: LancamentoFindWhere): Promise<LancamentoAutorizacaoHistory[]> {
+  async findMany(options?: FindManyOptions<LancamentoHistory> | undefined, andWhere?: LancamentoFindWhere): Promise<LancamentoHistory[]> {
     let whereCount = 0;
-    let qb = this.lancamentoAutorizacaoHistoryRepository
+    let qb = this.lancamentoHistoryRepository
       .createQueryBuilder('lancamentoHistory') //
       .leftJoinAndSelect('lancamentoHistory.autorizacoes', 'autorizacoes')
       .leftJoinAndSelect('lancamentoHistory.autor', 'autor')
@@ -124,11 +116,11 @@ export class LancamentoAutorizacaoHistoryRepository {
     return ret;
   }
 
-  getAll(): Promise<LancamentoAutorizacaoHistory[]> {
+  getAll(): Promise<LancamentoHistory[]> {
     return this.findMany();
   }
 
   async softDelete(id: number): Promise<DeleteResult> {
-    return await this.lancamentoAutorizacaoHistoryRepository.softDelete(id);
+    return await this.lancamentoHistoryRepository.softDelete(id);
   }
 }
